segmented filePathGen and Writing csv to file
